refactor: improve types and names of files
