+++ apps/client/src/app/pages/map/components/marker/components/MarkerDropDownMenu.tsx
@@ -1,11 +1,11 @@
 import { MutableRefObject } from 'react';
-import useRefMenu from '../hooks/useRefMenu';
+import useRefMenu from '../../../hooks/useRefMenu';
 
 function MarkerDropDownMenu({
   className = '',
   setIsOpen,
   children
-}: DropDownMenuProps) {
+}: MarkerDropDownMenuProps) {
   const ref = useRefMenu({ setIsOpen });
 
   return (
