+++ apps/client/src/app/pages/map/components/search-bar/components/menu/hooks/useMenuActions.tsx
@@ -7,7 +7,7 @@ function useMenuActions() {
 
   function goHelp() {}
 
-  const actions: MarkerMenuActions = {
+  const actions: MenuActions = {
     goHome,
     goProfile,
     goSettings,
