+++ apps/client/tsconfig.json
@@ -27,6 +27,9 @@
       "@/*": [
         "src/*"
       ],
+      "@src/*": [
+        "src/app/*"
+      ],
       "@assets/*": [
         "src/assets/*"
       ],
